refactor: support explicitly set BSON module
The BSON module is currently plucked off the passed in MongoClient
at the moment its needed. This can cause a number of problems, such
as using two different BSON modules (one for the AutoEncrypter one
for the key fetching client). We now support passing a single
module to the top level CSFLE classes, and that is used for all
BSON serialization.

NODE-2405